# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DFM Signage Hub API",
    "version": "0.2.28",
    "description": "Alle Routen liegen unter /api/v1. Anmeldung per Cookie (__Host-dfm_sid) mit CSRF-Header X-CSRF-Token bei schreibenden Aufrufen; Player nutzen ein Bearer-Token. WebSocket: siehe docs/websocket.md."
  },
  "servers": [
    {
      "url": "https://dfm-signage.local"
    }
  ],
  "paths": {
...
    "/api/v1/apps": {
      "get": {
        "summary": "Sitzung + Recht „devices.read“",
        "x-berechtigung": "devices.read",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Erfolg"
          },
          "400": {
            "description": "Eingabe ungültig (verständlicher Text im Feld \"error\")"
          },
          "401": {
            "description": "Nicht angemeldet"
          },
          "403": {
            "description": "Berechtigung fehlt"
          }
        }
      }
    },
...
    "/api/v1/apps/{type}": {
      "put": {
        "summary": "Sitzung + Recht „settings.manage“",
        "x-berechtigung": "settings.manage",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "enabled": {
                    "type": "boolean"
                  },
                  "config": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Erfolg"
          },
          "400": {
            "description": "Eingabe ungültig (verständlicher Text im Feld \"error\")"
          },
          "401": {
            "description": "Nicht angemeldet"
          },
          "403": {
            "description": "Berechtigung fehlt"
          }
        }
      }
    },
...
    "/api/v1/demo/remove": {
      "post": {
        "summary": "Sitzung + Recht „media.write“",
        "x-berechtigung": "media.write",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Erfolg"
          },
          "400": {
            "description": "Eingabe ungültig (verständlicher Text im Feld \"error\")"
          },
          "401": {
            "description": "Nicht angemeldet"
          },
          "403": {
            "description": "Berechtigung fehlt"
          }
        }
      }
    },
...
    "/api/v1/setup/admin": {
      "post": {
        "summary": "öffentlich (nur im Einrichtungs-/Pairing-Ablauf)",
        "x-berechtigung": null,
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code",
                  "name",
                  "password"
                ],
                "additionalProperties": false,
                "properties": {
                  "code": {
                    "type": "string",
                    "maxLength": 20
                  },
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 60
                  },
                  "password": {
                    "type": "string",
                    "maxLength": 300
                  },
                  "site": {
                    "type": "string",
                    "maxLength": 100
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Erfolg"
          },
          "400": {
            "description": "Eingabe ungültig (verständlicher Text im Feld \"error\")"
          }
        }
      }
    },
    "/api/v1/setup/state": {
      "get": {
        "summary": "öffentlich (nur im Einrichtungs-/Pairing-Ablauf)",
        "x-berechtigung": null,
        "parameters": [],
        "responses": {
          "200": {
            "description": "Erfolg"
          },
          "400": {
            "description": "Eingabe ungültig (verständlicher Text im Feld \"error\")"
          }
        }
      }
    },
...
    "/api/v1/system/hub": {
      "get": {
        "summary": "Sitzung + Recht „devices.read“",
        "x-berechtigung": "devices.read",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Erfolg"
          },
          "400": {
            "description": "Eingabe ungültig (verständlicher Text im Feld \"error\")"
          },
          "401": {
            "description": "Nicht angemeldet"
          },
          "403": {
            "description": "Berechtigung fehlt"
          }
        }
      }
    },
...
  }
}